Add permit(:role) to brakeman ignore file
Brakeman, the security scanner, raises a warning for mass assignment
when role is included in the allowed params. This happens in the
membership controller where we allow Organisation and Super admins to
update a users group role.

Brakeman issues a warning because allowing a User model to be updated
can be dangerous if there are no checks in place. In some applications
it may be possible to change a user's account into an admin.

In this case, we are allowing the role a user plays in an group to be
changed. This makes sense and isn't dangerous.