Call service method directly from controller
Ensure that a validation message is shown when no option is selected
on the "Make your form live" page. Previously, we just redirected back
to the form page with no message.

Remove the `make_live_input#make_form_live` method as it is redundant.
Just directly call the service method from the controller, as we have
already validated that the form input is valid.