Display correct error messages for "Make your form live"
Ensure that the correct translation is used for error messages when
tasks for the form are incomplete. Previously, the error message
displayed was "missing_pages" etc. This was acheived by passing a
symbol rather than a string for the error type in
`make_live_input#required_parts_of_form_completed`.