Make is_optional a radio button
DAC raised the way we were using the is_optional checkbox as confusing
and a violation of WCAG 2.2 3.3.2 Labels or Instructions. The logic for
this was that it wasn't clear whether the checkbox was mandatory or not.

To fix this, this commit changes the is_optional field into a binary
radio button, so it's clearer to the user what the two options are.

At present, all of the existing pages in forms-api have is_optional set
to either `true` or `nil`. We should migrate that database so that only
`true` or `false` are possible. In the meantime, we'll set the value to
`false` when editing an existing question where is_optional is `nil`.