Simplify boolean casting
Because the form in the view can only pass strings, not Booleans, we
have some logic in the page model to convert the string into a boolean
before validation, so that it can be sent to the API as a boolean.

By using `ActiveRecord::Type::Boolean.new.cast` we can avoid having to
maintain our own way of casting to Boolean. This allows us to simplify
some of the other logic around this a bit too. We may be able to
simplify this even more once we've migrated the API field to be
non-nullable.