Add default group associations
When adding a new form not in a group (legacy methods), we want to add
it to a group anyway, so we need a way to determine which group to add
it to. This commit adds a `default_for`/`default_group` association
between a groups and an organisation or trial user.