Add controller actions for reviewing upgrade requests
Add GET and POST actions for /groups/:id/review_upgrade.

We call different service methods depending on whether "Yes" or "No"
was selected on the confirmation page:

- If "Yes" is selected, the group status is upgraded to active and
group admins are emailed.
- If "No" is selected, the group status is upgraded to trial and group
admins are emailed.

We are using the `review_upgrade?` Pundit policy which validates the
user has permission to upgrade and that the group has status
"upgrade_requested".

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>