Fix bug with none of the above not being loaded
Fixes a small bug found while testing the change in the previous commit.

The is_optional field wasn't being passed to the SelectionSettingsInput
constructor, which meant that if the user:

- created a new selection question
- set 'include none of the above' to true
- progressed to the question form
- went back to the selection settings page

the value they had set for 'Include none of the above' would not be
preserved. We now pass the is_optional field to the constructor, so this
should now work as expected.