Remove redundant context in spec
Simplify the spec by removing the context for the user having
permission to upgrade a form but not having permission to request an
upgrade.

This tests in this context are identical to the the context where the
user has both permissions. The intention was to check that the
permission to upgrade a form took precedence when deciding the content
to display in the trial notification banner, but this can be achieved
by just having a single context where both permissions resolve to
true.