Add rake task to update the organisation for an MOU signature
Adds a rake task that, given the User name and Organisation name for
an MOU, will update the organisation for the MOU to the target
organisation.

The primary intended use for this is when a user and their forms are
moved to a sub-organisation and the MOU signature needs to be moved
with them to the sub-organisation to reflect the organisation that the
user agreed to the MOU for.