Update root path to depend on groups feature flag
Add a new HomepageController with an index action that is called when
the root path is visited.

If the groups feature flag is enabled, redirect to the groups index
page.

If the groups feature flag is disabled, directly call the Forms
controller index action. We're doing this rather than redirecting
because we don't want to add a new route for the Forms controller
index action, which will be removed soon anyway.