Update FeatureService to allow configuration per organisation
Allow specifying overrides for a feature flag per organisation by
providing an object in the settings YAML file as the value for the
feature, with optional keys matching the slugs for individual
organisations with underscores instead of dashes. If a key exists for
the organisation for the logged in user, the value of this key is used
over the value of the `enabled` key.

Example:

```
features:
  groups:
    enabled: false
    organisations:
      government_digital_service: true
```

This continues to support providing a simple boolean flag for a
feature.

We will also continue supporting the `enabled?` class method for the
FeatureService, which will ignore any overrides for the organisation
as it does not have access to the logged in user.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>