Remove unnecessary trait
The `:active` trait in the groups factory was configured to return
`:live`, which is not a valid form state. However, the tests were still
passing.

This is because it doesn't actually get used - [factorybot automatically
creates traits for enum attributes](https://thoughtbot.github.io/factory_bot/traits/enum.html#enum-traits),
so there was already an `:active` trait on the factory, which returns
the correct value `:active`.