Add new forms to organisation default group
When the groups feature is not enabled, all forms created by editors should be
added to a default group, owned by the organisation.

This is so that when the groups feature is enabled, these forms can be
seen in the groups view.

Each organisation will have a single "default group" which is created if
it does not already exist. The new group will have a status based on the
organisation's MOU signature. The user who is creating the form will be
added to this group as an editor, if they do not already belong to it.

Existing forms will be added to a default group with a rake task in the
future.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>