Add rake task to migrate forms into default groups
Add a new rake task to add a default group for each organisation.

Organisations which have at least one editor user are given a default
group. The status of the group is decided based on the organisations
MOU.

The organisation's editor users are added to the group. The
organisation's forms are added to this group, unless they belong to
other groups already.

To run the task:

```
bundle exec rake default_groups:create
```

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>