Add rake default_groups:create_dry_run
Wrap the rake task for migrating existing forms and editor users into default
groups in a transaction.

This prevents changes to the database being committed if an exception is
raised when running.

For example, the task will try to create a default group for any
organisation which doesn't have one. If that organisation already has a
group with same name, an error will be thrown and the task aborted. This
would leave the DB with some organisations having default groups and
some without.

Using a transaction means that either all the changes will happen, or
none of them.

This is used to create a dry_run version of the task. At the end of the
dry run task, a Rollback exception is raised. This means no changes are
committed to the database.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>