Change groups index to show "GOV.UK Forms team" if group has no creator
For groups created by the `default_groups:create` task (see #1213) we
don't set the `created_by` attribute. Currently this will show as
'Unknown' in the groups index for users, but for the default groups it
is more useful to say that the group was created by the GOV.UK Forms
team.

As groups created by any usual means through the application will have a
user assigned for `created_by`, so it's reasonable to just change the
`created_by_unknown` text for the `GroupList` component, rather than any
more complex mechanism.