Change trial_users:summary to show IDs for forms not in default group
Rather than showing the IDs for forms that are in any group, just show
the forms that are in a group that is not the default group.

This will now list forms that have been added to the organisation
default group previously, but not forms that have been added to the
trial user's default group previously.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>
Co-authored-by: Tom Iles <tom.iles@digital.cabinet-office.gov.uk>