Add google analytics behind flag
Adds a subset of the analytics code from the product pages repo to this
one and runs it when the settings flag is enabled.

We don't need most of the code that's in the product pages, because it's
an internal tool and we don't need to implement the cookie consent
mechanism.