Restrict analytics to non super admins
We're only really interested in what users from other teams are doing -
we don't really want to capture what members of our team are doing in
the platform using google analytics, because this may skew the results
and we have better ways of feeding this information back.

This commit makes it so theat we only initialise the analytics if the
settings flag is enabled, and the user is not a super admin.