Change tests to expect groups feature on
Change tests to expect the groups feature to be enabled.

Some tests have been wrapped in a context to turn the feature off,
where the feature does not make sense with groups enabled.

For example, the forms#index route will no longer be needed when groups
is on.

In other tests, which used root_path, have been changed to use
groups_path.