Change memberships table to be sorted by user name
Currently the order of users in the table on the "Members of this group"
page is whatever is returned by the database, which is not consistent;
so when editing users the table can jump around between page loads,
which is a bit disorientating.

This commit fixes things by ordering by name. This is a naive way, but
at least is more likely to have consistent results.