Fix warning about scoped order
Currently when running the `organisations:summary` task we get a warning
from Rails about

```
Scoped order is ignored, it's forced to be batch order.
```

This is because we're using `find_each`, but the
`Organisation.with_users` scope adds an `ORDER BY` clause.

This PR fixes things by removing the order relation with [unscope].

When testing locally the warning appears in `log/development.log`, but
in production it is emitted to stdout.

unscope: https://api.rubyonrails.org/classes/ActiveRecord/QueryMethods.html#method-i-order