Change Mailchimp synchronizer to archive users instead of permanent delete
Permanently deleting a user from a Mailchimp audience prevents them from
being added back [[1]], and seems to be mainly aimed at actioning GDPR
deletion requests [[2]].

In most cases where a user is being removed from a email subscription
list by our Rake task, it will be because we've denied them access to
the platform, but this could be a temporary state of affairs; for
instance if they haven't signed in for a while.

This commit changes `MailchimpListSynchronizer` so that it archives the
subscriber instead of permanently deleting them; they won't count
towards our billing but will be easy to add back if needed.

[1]: https://mailchimp.com/developer/marketing/api/list-members/delete-list-member/
[2]: https://mailchimp.com/help/delete-contacts/