Change forms:move to also update form organisation
When moving a form to a group in a different organisation to the one it
currently belongs to, we should update the form's organisation
attribute.

Not doing so shouldn't have any effect when the groups feature is enabled,
but until we remove the organisation attribute we should keep things
consistent.