Update FormTaskListService with new policies
Update the tasklist service to only consider groups.

Previously the tests created users as needed to test the tasklist
service. As the polices will be changed, the tests here also have to be
modified to reflect that changes.

As the tests for the tasklist are unit tests, not integration tests,
they shouldn't be testing behaviour within the polices.

The specs have been rewritten to stub policies which are tested in their
own unit tests.

This change should mean that future changes to the Form policy shouldn't
require changes to the tasklist tests too.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>