Add groups to request and feature specs
All feature and request specs should be arranged so that forms belong to
groups.

This allows the forms to be accessed by the tests with the groups
feature enabled.

Without this change, changing the FormPolicy to only consider groups
would mean that almost all tests fail due to permission issues.

The changes here are large, but mainly mechanical. Groups and
memberships are added to tests as needed.

A lot of tests don't care about the details of groups and users, only
that form exists which the user can access.

We considerd modifying factories to make groups and users as needed
instead of creating them in each test. We haven't made that change in
this commit, prefering instead to make the changes explicit before
cleaning them up.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>