Update the masquerading link text to show the user name too
With this change, we are updating the new masquerading link in the Users view
so that instead of repeating 'Masquerade as this user' in each row, it shows
the name of the individual users, when it is available.

If the name is missing, then the link text will be 'Masquerade as this user'
as before.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Marco Tranchino <marco.tranchino@digital.cabinet-office.gov.uk>