Remove the ability to masquerade while masquerading
With this change, we are preventing multiple masquerade actions which would
occur if a super admin masqueraded as another super admin, thus having the
option to masquerade while masquerading.

Since we only keep track of the previous user, multiple masquerading actions
would result in the user becoming a different super admin irreversibly.

By checking if the user is already masquerading, we can prevent a further
masquerade action from happening thus resolving the problem.

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Marco Tranchino <marco.tranchino@digital.cabinet-office.gov.uk>