Add setting to allow acting as user
With this first commit, we introduce a new setting for the new functionality
implemented with this and the following commits, for Firebreak July 2024.

When enabled, it will be possible for super admins to act (or masquerade) as
other users.

This functionality will be enabled on Staging and Dev only, to allow the
Product Team to view old and new features from the point of view of different
types of users.

Further information in Trello[1].

Designs in Mural[2].

[1]
https://trello.com/c/9abpfhCl

[2]
https://app.mural.co/t/gaap0347/m/gaap0347/1720002082246/bb065084ed274f561174c72964316652f4a4acd7

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Marco Tranchino <marco.tranchino@digital.cabinet-office.gov.uk>