Add act as user button to users list
We have decided to have an entry point for this new functionality just in the
existing Users page, for the time being.

Note that the Users page is only available to super admins.

The button to start masquerading and acting as another user is only displayed
for users who are not themselves super admins.

Further information in Trello[1].

Designs in Mural[2].

[1]
https://trello.com/c/9abpfhCl

[2]
https://app.mural.co/t/gaap0347/m/gaap0347/1720002082246/bb065084ed274f561174c72964316652f4a4acd7

Co-authored-by: Laurence de Bruxelles <laurence.debruxelles@digital.cabinet-office.gov.uk>
Co-authored-by: Samuel Culley <samuel.culley@digital.cabinet-office.gov.uk>
Co-authored-by: Marco Tranchino <marco.tranchino@digital.cabinet-office.gov.uk>