Remove EventLogger
It looks like EventLogger was added to mirror the EventLogger in
runner, which is used to log events in the form filler journey in a
consistent way.

Here, it was simply logging the provided hash as JSON. This doesn't
appear to offer an advantage over using `Rails.logger.info` which will
now append default attributes about the request context to all log
lines.

Remove EventLogger, and replace with logging more descriptive
messages and passing in the appropriate custom JSON attributes for the
log lines.