Make find invited user by email case insensitive
When users are created in Signon and Auth0, the email is downcased
when it is stored.

When inviting users to join groups, we should therefore ignore the
case when locating the user to invite. Previously, we were showing an
error saying that the user does not have a Forms account when the
case did not match the stored email for the user.