Fix rendering act as user banner on error pages
When an error page is rendered by the ErrorsController, the current_user
instance variable is not set. This commit changes the application layout
to use the new act as user view helpers, instead of relying on the
current user.

This means that users who are acting as another and reach an error state
will still be able to stop acting as another and recover.