Configure custom logger for application logs
We already use Lograge to add attributes about the request context,
such as the `request_id`, `user_id` etc to the request logs for
incoming requests to the app.

Configure the app to use the new `ApplicationLogger` class which makes
use of the `CurrentAttributes` implementation to extract common
logging fields and appends them to the JSON object for a log event.

The Lograge configuration has also been updated to extract these
fields from the `CurrentLoggingAttributes` so we get the same set of
fields on both the application and request logs.