Make logs less noisy for local development
Configure the Rails default logger for application logs, and configure
Lograge so that it doesn't add our custom logging context when
running locally.

The intention of this is to make logs easier to parse when developing
locally. If we want to test the production output, we can comment out
these lines in development.rb