Add 'pending' to component test template
By default, the test file generated by the custom component generator
will violate rubocop's RSpec/EmptyExampleGroup rule, because there are
no examples in the file.

This commit adds a `pending` statement to the file, which satisfies the
rubocop rule and means that the user will get a non-failing warning
message when running the test suite until they replace this statement
with some tests.