Revert change to use rails default logger when running locally
Our custom ApplicationLogger has a slightly different interface than
the rails default logger which allows for passing in a hash of
additional arguments to add to the structured log as a second
parameter.

This means we can't just swap this out for the ActiveSupport::Logger
class when running in development mode.

Undo the change to make logs less noisy for development by swapping
the logger. If we want to do this, we will have to make changes to
the ApplicationLogger to allow for this.