Revert change to make local request logs less noisy
This change was made with the intention of making logs less noisy for
local development by removing our custom log context (request_id,
user_id etc) from the logs.

Initially we did this for the application (non-request) logs too, but
this change broke the logging locally so it was reverted.

Also revert the change for the request logs so that application and
request logs are consistent. We can revisit this if we find that logs
are too noisy for local development.