Add stub to MailchimpListSynchronizer specs
The tests for the MailchimpListSynchronizer print output when they are
run.

This commit stubs puts in the the MailchimpListSynchronizer tests to
stop this output cluttering the terminal when they are run.

I think there are other ways to do this globally for all tess. I've gone
for stubbing puts because it keeps the solution close to the cause and
  leaves the option of using puts in specs elsewhere if needed.