Reduce authorisation session timeout to 8 hours
Change the length of time the authorisation sesssion is valid for to
8 hours.

This will force a check with Auth0 that the session with
Auth0 is still valid, if it is the user won't be required to login
again. We will separately adjust the Auth0 tenant configuration
to reduce the length of time the Auth0 session is valid for.

We are reducing this to 8 hours due to the following points in our
risk treatment plan:

RTP ref 7:

For Super Administrators accessing the Forms Administration Interface,
 implement the shortest tolerable idle session timeout and absolute
 session timeout.

RTP ref 3:

For Editors accessing the Forms Administration Interface, implement
the shortest tolerable idle session timeout and absolute session
timeout.

We have determined that 8 hours is the shortest tolerable time when
balancing user experience as we expect users to be able to user the
platform for a standard working day without interruption.