Remove role change check in user organisation validation
This commit removes the check for role change to editor, because the
editor role is going to removed. In future, all trial and editor users
will be migrated to standard users, and new users will have the
standard role by default.

We retain the check for role change to
organisation_admin. This is because it isn't valid to be
organisation_admin without an organisation.

Co-authored-by: Stephen Daly <stephen.daly@digital.cabinet-office.gov.uk>