Add spec for when user chooses not to delete form
We weren't testing the path when a user presses the delete button for a
form but then chooses not to delete it at the confirmation page.

Not testing this path led to a bug being introduced in commit 5a05d268
as we were asking Pundit to verify that we authorize the user for all
actions, but weren't testing all paths for all actions.