Move authorization for delete_confirmtion_controller#destroy to start
Because we were authorizing the delete in the `delete_page` and
`delete_form` helpers, we weren't checking the user was authorized when
they decided not to delete or did not send a valid request.

This commit fixes things by making sure that checking the user's
authorization is the first things we do in the action.