Make sure groups feature is enabled everywhere
We're confident now that we want to use the groups feature going
forward, and have started removing the legacy code, so we want it
enabled everywhere.

This commit removes the groups feature flag override from settings/*.yml
files so that the global value is used and we have the groups feature
enabled for everyone everywhere.