Attempt to fix flaky tests
These tests were checking that 2 arrays of groups/forms were exactly
equal. However, as the code under test was retreiving the groups/forms
from the database and not applying sorting, the order was
unpredictable. This was causing the tests to occasionally fail.

Replace `eq` assertions with `match_array`, which ignores the ordering

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>