Replace usages of `trial_user` helper with `standard_user`
Before we migrate trial and editor users to the standard role, we want
to have confidence that the standard role is behaving as expected.
Replace tests that use a trial user with using a standard user.

We have already made sure that trial, editor and standard users are
treated the same in the application code.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>