Display standard user role option on users screen correctly
Add translations to display the radio button for selecting the
standard user rule on the edit user page correctly.

For now, it will still be possible to select "Trial" and "Editor"
roles, but we will remove these options after we have migrated trial
and editor users to be standard users.