Remove `editor_user` factory
Remove the `editor_user` factory. Instead, default to creating
standard users wherever possible. In the cases where we still want
to create an editor user where we want to retain tests until editor
users are removed, use the dafault `editor` factory that is added
automatically for the `User.role` enum.

Co-authored-by: Jamie Wilkinson <jamie.wilkinson@digital.cabinet-office.gov.uk>