Add users_reports_service
Add a new service for returning data about organisations and the users
they have.

The service has one method, user_data, which returns a hash which can be
passed to a govuk_table component.

This hash could also be used to create CSV or other formats as it has a
caption, headings and rows.

The query itself produces an array of arrays with organisation name and
the number of users they have.

For an example, [["Government Digital Service", 112], ["Department for
testing", 12], [nil, 112]]

Users without an organisation are counted as [nil, count]. Organisations
without any users will not be included in the results.

The query is a bit hairy because we are being careful to only select the
column we need from the database, not whole models. The results are
ordered by count, then organisation name. The organisation with the most
users will appear first in the array followed by the others in
descending order.

The results of this query are converted into the format for
govuk_tables. The existing translation for users without an organisation
is used.

The tests for this are terse. This is partly because we can compare the contents
of the hash directly against the expected output. This is effective and
covers the behavior but it doesn't document everything very clearly.
For example the order is tested but it's not totally obvious.

I think it makes sense to keep the tests like this given the
functionality of the feature.