Make DraftQuestion forwards compatible with Page response
Currently if an attribute is added to the Page model in forms-api trying
to edit the question for that page will result in an error:

     Failure/Error: edit_draft_question.attributes = page.attributes.except(:id, :position, :next_page, :has_routing_errors, :routing_conditions, :question_with_text)

     ActiveModel::UnknownAttributeError:
       unknown attribute 'newly_added_to_api' for DraftQuestion.
     # ./app/controllers/pages_controller.rb:65:in `setup_draft_question_for_existing_page'
     # ./app/controllers/pages_controller.rb:47:in `draft_question'
     # ./app/controllers/pages/questions_controller.rb:3:in `block in <class:QuestionsController>'
     # ./spec/requests/pages/questions_controller_spec.rb:171:in `block (4 levels) in <top (required)>'
     # ./spec/support/features.rb:19:in `block (2 levels) in <top (required)>'

This is happening because the attributes passed to `edit_draft_question`
includes the new attributes, but these attributes are not in the schema
for the `draft_questions` table.

This commit fixes things by filtering the attributes set in
`edit_draft_question` to only those known to the `DraftQuestion` model.