Allow archived users to be resubscribed
Because our API call fetches users with all statuses, `existing_members`
included archived members.

This meant that they were counted as active members and as a result
would not be resubscribed if their Forms accounts were reactivated.

This commit filters the results we get back from the API, so that only
subscribed users are counted and users with other statuses can be
resubscribed. It also includes:

- renaming `existing_members` to `currently_subscribed_members` for
clarity
- changing the `set_list_member` API call so that it will set the status
for existing records