Improve negative test reliability
The shared example we had previously for testing that the API didn't
receive a subscribe or delete request was a bit risky. Since we were
using `not_to` against a very specific request, a slight change to the
result might result in a false negative.

This commit updates the test to be more general - instead of using
`not_to` as we were, we now test that the API only receives the request
once, with a different email address.