Allow passing existing user to login_as_super_admin_user
For security reasons super admin users need to sign in in a slightly
different way in our app, including in our specs. Currently our spec
helpers only make it easy to sign in as the default super admin user
before a request spec, this commit adds to the helper for signing in as
a super admin an optional argument that supplies an existing user.