Prevent link navigation for jsdom
We do most of our tests in the jsdom environment, which describes itself
as 'a pure-JavaScript implementation of many web standards'. One of the
web standards it does _not_ implement is navigation between pages. This
causes it to output errors if tests try to navigate between pages (e.g.
by clicking a link, as we do in the google tag tests).

This PR removes the default click event behaviour from the link we click
in the google tag tests. This means that the click event in the test
will not try to navigate between pages, so jsdom will not throw an
error.

We could also have resolved this error by:

- using an alternative environment to jsdom that supports navigation. I
  decided against this since the test in question doesn't actually care
  about navigation, so stubbing out that behaviour makes more sense.
- silencing the error - this felt like a bad idea, since it runs the
  risk of silencing useful errors, and it's cleaner to stub out the
  behaviour we don't want.