Add repeatable_page_enabled feature flag
Add a new feature flag, repeatable_page_enabled, which will be used to
decide whether to allow pages to be set as is_repeatable or not.

This PR only adds the setting. It doesn't do anything with it.

To turn the feature on, either set it in your environment's settings
file or as a shell environment variable:

```
SETTINGS__FEATURES__REPEATABLE_PAGE_ENABLED=true bundle exec rails s
```