Change page model to convert is_repeatable to bool
This ensures that all requests sent to the API use a json boolean value
rather than text.

This shouldn't matter to the API which should cast "false" as false
anyway.

Doing it here makes testing the API json requests easier and is slightly
saner.

It's done in the Page model rather than in the DraftQuestion model
because DraftQuestion can pass nil values when there is no change to
is_repeatable.

It feels like a clunky way to do things but I think that reflects
ActiveResource, which doesn't support the Attribute API. Instead it uses
it's own schema.