Add is_repeatable to page controller
The is_repeatable boolean value for pages is added to the controller.

The controller allows the value to be set via param regardless of
whether the feature flag it set.

The tests check that the value is preserved if the params do not
include the value. This would happen when the feature flag is off and
the form input isn't included in the page.